requirements.txt
python-dotenv==1.0.0
Flask==3.0.3
waitress==3.0.0
